Implement some group-related stuff wrt the database
I'm definitely missing a bunch of important functionality and data
when it comes to what the groups are actually going to do but maybe this
is a good start? Or possibly everything is terrible and I'll have to
redo all the code, but let's try to think positively.

Also the front-end stuff that you see in the test activity is absolute
garbage, take it with a grain of salt since I'm mostly just using it as
a placeholder to test the internal logic rather than any sort of proof
of concept.