Making changes to how users are organized
Split up users that are currently "online" (with location info) vs.
permanent user information (like name, profile pic). I /think/ this will
be more efficient in the end? (The problem is you have to use a lot of
listeners and stuff which is kind of annoying q_q Hopefully it all works
out!!!!)